Initial work to convert xspress-detector for using fastCS
In this commit I added my initial work to make this run agains the FastCS framework.
Managed to use AsyncParameterTree and update get method so that it would return elements on the list.
Currently only working with the config parameters but will add others when the code is more robust.
Will remove bits of commented code in the future.